pyproject.toml
[project]
name = "ppo_planer"
version = "0.1.0"
description = "PPO Planning project"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.1.0",
    "ale-py==0.8.1",
    "certifi==2024.2.2",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cloudpickle==3.0.0",
    "colorama==0.4.6",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "farama-notifications==0.0.4",
    "filelock==3.13.1",
    "fonttools==4.51.0",
    "fsspec==2024.2.0",
    "grpcio==1.62.2",
    "gymnasium==0.29.1",
    "idna==3.7",
    "importlib-resources==6.4.0",
    "jinja2==3.1.3",
    "kiwisolver==1.4.5",
    "markdown==3.6",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.8.4",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "networkx==3.2.1",
    "numpy==1.26.3",
    "opencv-python==4.9.0.80",
    "packaging==24.0",
    "pandas==2.2.2",
    "pillow==10.2.0",
    "protobuf==5.26.1",
    "psutil==5.9.8",
    "pygame==2.5.2",
    "pygments==2.17.2",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.1",
    "requests==2.31.0",
    "rich==13.7.1",
    "sb3-contrib==2.3.0",
    "scipy==1.13.0",
    "shimmy==1.3.0",
    "six==1.16.0",
    "stable-baselines3==2.3.1",
    "sympy==1.12",
    "tensorboard==2.16.2",
    "tensorboard-data-server==0.7.2",
    "torch==2.3.0",
    "torchaudio==2.3.0",
    "torchvision==0.18.0",
    "tqdm==4.66.2",
    "typing-extensions==4.9.0",
    "tzdata==2024.1",
    "urllib3==2.2.1",
    "werkzeug==3.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
